sdr_backend/requirements.txt
# Web framework and server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
httpx>=0.27.0
tenacity==8.2.3

# Document processing
# pytesseract==0.3.10
# Pillow==9.5.0  # Last version supporting Python 3.8
# pdfplumber==0.10.3
# pdf2image==1.17.0

# AI/ML
# langchain>=0.1.0,<0.2.0
# langchain-openai>=0.0.5
# langchain-community>=0.0.19
openai==1.75.0
anthropic==0.49.0
torch>=2.3.0
--find-links https://download.pytorch.org/whl/torch_stable.html
transformers>=4.41.0
huggingface-hub>=0.20.0
faiss-cpu==1.7.4
scikit-learn==1.2.2
accelerate==0.26.1
numpy>=1.19.0,<2.0.0

# Database
sqlalchemy>=2.0.7
greenlet==3.2.3
psycopg2-binary==2.9.9
pandas==2.0.3  # Last version supporting Python 3.8
# aiosqlite==0.19.0
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
databases==0.9.0
alembic==1.13.1
supabase==2.15.0  # Python client for Supabase
python-jose==3.3.0  # Often needed with Supabase for JWT handling
postgrest>=0.10.6  # Dependency for Supabase

# Utils
python-dotenv==1.0.1
pydantic==2.11.3
pydantic-settings==2.1.0
pyyaml==6.0.1
PyJWT==2.8.0
sentence-transformers==3.2.1

#Threat Modelign DFD
pytm==1.3.1
# graphviz==0.20.3

prometheus_fastapi_instrumentator==5.11.2
d2py==0.7.0
# (Optional) If you later install a working Python binding for ELK, add
# it here.  The current PyPI "elk" build fails on Python ≥3.12, so we
# keep it out of the locked requirements to avoid CI install errors.

# Testing
pytest>=8.0.0

# Rate limiting
slowapi==0.1.9

beautifulsoup4==4.13.4
lxml==4.9.3
rapidfuzz==3.13.0
symspellpy==6.9.0